\name{testarma2-package}
\alias{testarma2-package}
\alias{testarma2}
\docType{package}
\title{
\packageTitle{testarma2}
}
\description{
\packageDescription{testarma2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testarma2}
\packageIndices{testarma2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{testarma2}

Maintainer: \packageMaintainer{testarma2}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
